Blindaje de la aplicacion usando una politica

+++ app/Http/Controllers/UserController.php
@@ -12,11 +12,14 @@ use App\User;
 class UserController extends Controller
 {
     
-
+    public function __construct(){
+        $this->middleware('auth');
+    }
 
     public function index()
     {
-        //Gate::authorize('haveaccess','role.index');
+        //Gate::authorize('haveaccess','user.index');
+        $this->authorize('haveaccess','user.index');
         $users = User::with('roles')->orderBy('id', 'Desc')->paginate(2);
         return view('user.index', compact('users'));
         
@@ -25,6 +28,7 @@ class UserController extends Controller
 
     public function show(User $user)
     {
+        $this->authorize('view', [$user, ['user.show','userown.show']]);
         $roles = Role::orderBy('id')->get();
         return view('user.show', compact('user','roles'));
     }
@@ -33,6 +37,8 @@ class UserController extends Controller
     public function edit(user $user)
     {
         //$this->authorize('haveaccess','role.edit');
+        $this->authorize('update', [$user, ['user.edit','userown.edit']]);
+        
         $roles = Role::orderBy('id')->get();
         return view('user.edit', compact('roles','user'));
     }
@@ -41,6 +47,7 @@ class UserController extends Controller
     public function update(Request $request, User $user)
     {
         //$this->authorize('haveaccess','role.edit');
+        
 
         //validation
         $request->validate([
@@ -63,6 +70,8 @@ class UserController extends Controller
     public function destroy(User $user)
     {
         //$this->authorize('haveaccess','user.destroy');
+        $this->authorize('haveaccess','user.destroy');
+        
         $user->delete();
         return redirect()->route('user.index')->with('status_success', 'User successfully removed');
     }
